requirements.txt
scapy>=2.5.0
pandas>=2.0.0
matplotlib>=3.7.0
pyyaml>=6.0
rich>=13.0.0
